organize TreeViewNode utils
