+++ dbux-code/src/applicationsView/ApplicationNodeProvider.js
@@ -1,6 +1,6 @@
 import allApplications from '@dbux/data/src/applications/allApplications';
-import BaseTreeViewNodeProvider from '../codeUtil/BaseTreeViewNodeProvider';
-import EmptyTreeViewNode from '../codeUtil/EmptyTreeViewNode';
+import BaseTreeViewNodeProvider from '../codeUtil/treeView/BaseTreeViewNodeProvider';
+import EmptyTreeViewNode from '../codeUtil/treeView/EmptyNode';
 import ApplicationNode from './ApplicationNode';
 
 export default class ApplicationNodeProvider extends BaseTreeViewNodeProvider {
