+++ dbux-code/src/callGraphView/ContextNode.js
@@ -2,7 +2,7 @@ import { makeContextLabel } from '@dbux/data/src/helpers/makeLabels';
 import allApplications from '@dbux/data/src/applications/allApplications';
 import ExecutionContext from '@dbux/common/src/types/ExecutionContext';
 import traceSelection from '@dbux/data/src/traceSelection';
-import BaseTreeViewNode from '../codeUtil/BaseTreeViewNode';
+import BaseTreeViewNode from '../codeUtil/treeView/BaseTreeViewNode';
 
 export default class ContextNode extends BaseTreeViewNode {
   /**
