+++ dbux-code/src/callGraphView/ErrorNode.js
@@ -1,6 +1,6 @@
 import { makeTraceLabel } from '@dbux/data/src/helpers/makeLabels';
 import traceSelection from '@dbux/data/src/traceSelection';
-import BaseTreeViewNode from '../codeUtil/BaseTreeViewNode';
+import BaseTreeViewNode from '../codeUtil/treeView/BaseTreeViewNode';
 
 export default class ErrorNode extends BaseTreeViewNode {
   /**
