+++ dbux-code/src/codeUtil/treeView/BaseTreeViewNodeProvider.js
@@ -3,9 +3,9 @@ import EmptyObject from '@dbux/common/src/util/EmptyObject';
 import { newLogger } from '@dbux/common/src/log/logger';
 import NestedError from '@dbux/common/src/NestedError';
 import { makeDebounce } from '@dbux/common/src/util/scheduling';
-import { getThemeResourcePath } from './codePath';
-import { registerCommand } from '../commands/commandUtil';
-import { emitTreeViewAction, emitTreeViewCollapseChangeAction } from '../userEvents';
+import { getThemeResourcePath } from '../codePath';
+import { registerCommand } from '../../commands/commandUtil';
+import { emitTreeViewAction, emitTreeViewCollapseChangeAction } from '../../userEvents';
 import BaseTreeViewNode from './BaseTreeViewNode';
 
 /** @typedef { import("./BaseTreeViewNode").default } BaseTreeViewNode */
