+++ dbux-code/src/dataFlowView/DataFlowNodeProvider.js
@@ -3,8 +3,8 @@ import DataNodeType from '@dbux/common/src/types/constants/DataNodeType';
 import { newLogger } from '@dbux/common/src/log/logger';
 import allApplications from '@dbux/data/src/applications/allApplications';
 import traceSelection from '@dbux/data/src/traceSelection';
-import BaseTreeViewNodeProvider from '../codeUtil/BaseTreeViewNodeProvider';
-import EmptyTreeViewNode from '../codeUtil/EmptyTreeViewNode';
+import BaseTreeViewNodeProvider from '../codeUtil/treeView/BaseTreeViewNodeProvider';
+import EmptyTreeViewNode from '../codeUtil/treeView/EmptyNode';
 import DataFlowSearchModeType from './DataFlowSearchModeType';
 import DataFlowFilterModeType from './DataFlowFilterModeType';
 import ParentDataNode from './ParentDataNode';
