+++ dbux-code/src/editorTracesView/EditorTracesNodeProvider.js
@@ -1,7 +1,7 @@
 import groupBy from 'lodash/groupBy';
 import allApplications from '@dbux/data/src/applications/allApplications';
 import EmptyArray from '@dbux/common/src/util/EmptyArray';
-import BaseTreeViewNodeProvider from '../codeUtil/BaseTreeViewNodeProvider';
+import BaseTreeViewNodeProvider from '../codeUtil/treeView/BaseTreeViewNodeProvider';
 import { getTracesAt } from '../helpers/codeRangeQueries';
 import { getCursorLocation } from '../codeUtil/codeNav';
 import TraceNode from './nodes/TraceNode';
