+++ dbux-code/src/traceDetailsView/nodes/NavigationNode.js
@@ -2,7 +2,7 @@ import { window } from 'vscode';
 import tracePlayback from '@dbux/data/src/playback/tracePlayback';
 import traceSelection from '@dbux/data/src/traceSelection';
 import { newLogger } from '@dbux/common/src/log/logger';
-import BaseTreeViewNode from '../../codeUtil/BaseTreeViewNode';
+import BaseTreeViewNode from '../../codeUtil/treeView/BaseTreeViewNode';
 import { emitNavigationAction } from '../../userEvents';
 import { showInformationMessage } from '../../codeUtil/codeModals';
