+++ dbux-code/src/traceDetailsView/nodes/ObjectNode.js
@@ -2,7 +2,7 @@ import allApplications from '@dbux/data/src/applications/allApplications';
 import UserActionType from '@dbux/data/src/pathways/UserActionType';
 import traceSelection from '@dbux/data/src/traceSelection';
 import { emitSelectTraceAction } from '../../userEvents';
-import TraceNode from './TraceNode';
+import TraceNode from '../../codeUtil/treeView/TraceNode';
 
 export default class ObjectNode extends TraceNode {
   get clickUserActionType() {
