+++ dbux-code/src/traceDetailsView/nodes/TraceValueNode.js
@@ -2,7 +2,7 @@ import traceSelection from '@dbux/data/src/traceSelection';
 import allApplications from '@dbux/data/src/applications/allApplications';
 import { makeTraceValueLabel } from '@dbux/data/src/helpers/makeLabels';
 import { valueRender } from '../valueRender';
-import TraceNode from './TraceNode';
+import TraceNode from '../../codeUtil/treeView/TraceNode';
 
 /**
  * Children of `NearbyValuesTDNode`
