+++ samples/__samplesInput__/nested-try-finally.js
